solve circuitur import
